Refactoring
- move each setting row into its own file
- use gsetting specific names, since we will need setting rows independant of gsetting
