+++ lib/view/pages/sound/sound_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
-import 'package:settings/view/widgets/settings_row.dart';
 import 'package:settings/view/widgets/settings_section.dart';
+import 'package:settings/view/widgets/single_gsetting_row.dart';
 
 class SoundPage extends StatelessWidget {
   const SoundPage({Key? key}) : super(key: key);
@@ -10,17 +10,17 @@ class SoundPage extends StatelessWidget {
     return Column(
       children: const [
         SettingsSection(headline: 'System', children: [
-          BoolSettingsRow(
+          SingleGsettingRow(
             actionLabel: 'Allow Volume above 100%',
             settingsKey: 'allow-volume-above-100-percent',
             schemaId: 'org.gnome.desktop.sound',
           ),
-          BoolSettingsRow(
+          SingleGsettingRow(
             actionLabel: 'Event sounds',
             settingsKey: 'event-sounds',
             schemaId: 'org.gnome.desktop.sound',
           ),
-          BoolSettingsRow(
+          SingleGsettingRow(
             actionLabel: 'Input feedback sounds',
             settingsKey: 'input-feedback-sounds',
             schemaId: 'org.gnome.desktop.sound',
